+++ src/commands/index.ts
@@ -1,5 +1,6 @@
 import { ExtensionContext } from "vscode";
 import { registerAuthCommands } from "./auth";
+import { registerCommentCommands } from "./comments";
 import { registerEditorCommands } from "./editor";
 import { registerFileCommands } from "./file";
 import { registerFollowCommands } from "./follow";
@@ -7,6 +8,7 @@ import { registerGistCommands } from "./gist";
 
 export function registerCommands(context: ExtensionContext) {
   registerAuthCommands(context);
+  registerCommentCommands(context);
   registerEditorCommands(context);
   registerFollowCommands(context);
   registerGistCommands(context);