+++ src/extension.ts
@@ -1,5 +1,6 @@
 import { ExtensionContext } from "vscode";
 import { registerCommands } from "./commands";
+import { registerCommentController } from "./comments";
 import { registerFileSystemProvider } from "./fileSystem";
 import { store } from "./store";
 import { initializeAuth } from "./store/auth";
@@ -10,6 +11,7 @@ export function activate(context: ExtensionContext) {
   registerCommands(context);
   registerFileSystemProvider(store);
   registerTreeProvider(store, context.extensionPath);
+  registerCommentController();
 
   initializeStorage(context);
   initializeAuth();